Add rest of checks, remove old checks
